Add confirm dialog (#256)
* add confirmation dialog on delete tokens

* sort empty token categories to back

* add catch block

* fix focus

* refactor deleting token set to confirm dialog

* refactor clearTokens to confirmDialog

* bump version

+++ src/app/components/TokenGroupHeading.tsx
@@ -3,13 +3,21 @@ import {ContextMenu, ContextMenuTrigger, MenuItem} from 'react-contextmenu';
 import {useDispatch, useSelector} from 'react-redux';
 import Heading from './Heading';
 import {Dispatch, RootState} from '../store';
+import useConfirm from '../hooks/useConfirm';
 
 export default function TokenGroupHeading({label, path, id}) {
     const {activeTokenSet, editProhibited} = useSelector((state: RootState) => state.tokenState);
     const {deleteTokenGroup} = useDispatch<Dispatch>().tokenState;
+    const {confirm} = useConfirm();
 
-    const onDelete = () => {
-        deleteTokenGroup({parent: activeTokenSet, path});
+    const onDelete = async () => {
+        const userConfirmation = await confirm({
+            text: 'Delete group?',
+            description: 'Are you sure you want to delete this group?',
+        });
+        if (userConfirmation) {
+            deleteTokenGroup({parent: activeTokenSet, path});
+        }
     };
 
     return (
